java: Fixed precision loss for the unsigned max value for the primitive types.

+++ pymavlink/generator/java/lib/Messages/MAVLinkPayload.java
@@ -11,13 +11,13 @@ import java.nio.ByteBuffer;
 public class MAVLinkPayload {
 
     private static final byte UNSIGNED_BYTE_MIN_VALUE = 0;
-    private static final byte UNSIGNED_BYTE_MAX_VALUE = Byte.MAX_VALUE - Byte.MIN_VALUE;
+    private static final short UNSIGNED_BYTE_MAX_VALUE = Byte.MAX_VALUE - Byte.MIN_VALUE;
 
     private static final short UNSIGNED_SHORT_MIN_VALUE = 0;
-    private static final short UNSIGNED_SHORT_MAX_VALUE = Short.MAX_VALUE - Short.MIN_VALUE;
+    private static final int UNSIGNED_SHORT_MAX_VALUE = Short.MAX_VALUE - Short.MIN_VALUE;
 
     private static final int UNSIGNED_INT_MIN_VALUE = 0;
-    private static final int UNSIGNED_INT_MAX_VALUE = Int.MAX_VALUE - Int.MIN_VALUE;
+    private static final long UNSIGNED_INT_MAX_VALUE = Integer.MAX_VALUE - Integer.MIN_VALUE;
 
     private static final long UNSIGNED_LONG_MIN_VALUE = 0;
